Refactors Finders.
Refactors functions which find elements by label. Since the `for`
attriubte is looking for a corresponding `id` we don't need separate
functions for finding `input`, `select`, etc. by label.

* Removes `findInputByLabel` in favor of `findByLabel`
* Organizes finder functions pertaining to form elements in a
corresponding file.
