+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## 0.0.3
+* Removes `findInputByLabel` in favor of `findByLabel`
+* Nests finders pertaining to form elements under `helpers/finder`
