+++ tests/dummy/app/templates/application.hbs
@@ -1,4 +1,5 @@
 {{outlet}}
+{{partial "elements/select"}}
 {{partial "elements/article"}}
 {{partial "elements/aside"}}
 {{partial "elements/checkbox"}}
@@ -11,7 +12,6 @@
 {{partial "elements/ol"}}
 {{partial "elements/p"}}
 {{partial "elements/section"}}
-{{partial "elements/select"}}
 {{partial "elements/table"}}
 {{partial "elements/ul"}}
 {{link-to 'First link' 'firstLink'}}
