+++ tests/dummy/app/templates/elements/select.hbs
@@ -5,3 +5,10 @@
   <option value="value2">Value 2</option>
   <option value="value3">Value 3</option>
 </select>
+
+<label for="select_2">Label for second select</label>
+<select id="select_2" name="select_2">
+  <option value="value4" selected>Value 4</option>
+  <option value="value5">Value 5</option>
+  <option value="value6">Value 6</option>
+</select>
